removed CircleAstrologyIcons from page
